# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PWR Risk Review Coordinator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /meetings/{id}:
    put:
      operationId: updateMeeting
      tags: [meetings]
      summary: Update a meeting
      parameters:
        - $ref: "#/components/parameters/MeetingId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetingUpdate"
      responses:
        "200":
          description: Updated meeting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /email-drafts/{id}:
    put:
      operationId: updateEmailDraft
      tags: [email-drafts]
      summary: Update an email draft
      parameters:
        - $ref: "#/components/parameters/EmailDraftId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailDraftUpdate"
      responses:
        "200":
          description: Updated draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailDraft"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
...
    MeetingId:
      name: id
      in: path
      required: true
      schema:
        type: integer
    EmailDraftId:
      name: id
      in: path
      required: true
      schema:
        type: integer
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Meeting:
      type: object
      required:
        - id
        - requestId
        - meetingType
        - status
        - rescheduledCount
      properties:
        id:
          type: integer
        requestId:
          type: integer
        meetingType:
          type: string
        targetDate:
          type: ["string", "null"]
        scheduledStart:
          type: ["string", "null"]
        scheduledEnd:
          type: ["string", "null"]
        timezone:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        body:
          type: ["string", "null"]
        teamsLink:
          type: ["string", "null"]
        outlookEventId:
          type: ["string", "null"]
        status:
          type: string
        riskLead:
          type: ["string", "null"]
        rescheduledCount:
          type: integer
        notes:
          type: ["string", "null"]
...
    MeetingUpdate:
      type: object
      properties:
        meetingType:
          type: string
        targetDate:
          type: ["string", "null"]
        scheduledStart:
          type: ["string", "null"]
        scheduledEnd:
          type: ["string", "null"]
        timezone:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        body:
          type: ["string", "null"]
        teamsLink:
          type: ["string", "null"]
        status:
          type: string
        riskLead:
          type: ["string", "null"]
        rescheduledCount:
          type: integer
        notes:
          type: ["string", "null"]
...
    EmailDraft:
      type: object
      required:
        - id
        - requestId
        - subject
        - body
        - status
      properties:
        id:
          type: integer
        requestId:
          type: integer
        meetingId:
          type: ["integer", "null"]
        templateId:
          type: ["integer", "null"]
        templateType:
          type: ["string", "null"]
        toRecipients:
          type: string
        ccRecipients:
          type: string
        subject:
          type: string
        body:
          type: string
        status:
          type: string
        createdAt:
          type: ["string", "null"]
        updatedAt:
          type: ["string", "null"]
        sentAt:
          type: ["string", "null"]
    EmailDraftUpdate:
      type: object
      properties:
        toRecipients:
          type: string
        ccRecipients:
          type: string
        subject:
          type: string
        body:
          type: string
        status:
          type: string
...